Return the polymod-average population age distribution in 5y

returns the polymod-average population age distribution in 5y increments (weight country population distributions by number of participants). Note that we don't want to weight by survey age distributions for this, since the total number of participants represents the sampling. It uses the participant data from the polymod survey as well as the age specific population data from socialmixr R package to return the age specific average population of different, countries weighted by the number of participants from those countries who participated in the polymod survey.

get_polymod_population(
  countries = c("Belgium", "Finland", "Germany", "Italy", "Luxembourg", "Netherlands",
    "Poland", "United Kingdom")
)

Arguments

countries

countries to extract data from. Default is to get: Belgium, Finland, Germany, Italy, Luxembourg, Netherlands, Poland, and United Kingdom.

Value

data frame with two columns: lower.age.limit and population

Examples

get_polymod_population()
#> # A tibble: 21 × 2
#>    lower.age.limit population
#>              <int>      <dbl>
#>  1               0   1852682.
#>  2               5   1968449.
#>  3              10   2138897.
#>  4              15   2312032.
#>  5              20   2407486.
#>  6              25   2423602.
#>  7              30   2585137.
#>  8              35   2969393.
#>  9              40   3041663.
#> 10              45   2809154.
#> # … with 11 more rows
get_polymod_population("Belgium") 
#> # A tibble: 21 × 2
#>    lower.age.limit population
#>              <int>      <dbl>
#>  1               0     583492
#>  2               5     593148
#>  3              10     632157
#>  4              15     626921
#>  5              20     649588
#>  6              25     663176
#>  7              30     705878
#>  8              35     773177
#>  9              40     823305
#> 10              45     779436
#> # … with 11 more rows
get_polymod_population("United Kingdom") 
#> # A tibble: 21 × 2
#>    lower.age.limit population
#>              <int>      <dbl>
#>  1               0    3453670
#>  2               5    3558887
#>  3              10    3826567
#>  4              15    3960166
#>  5              20    3906577
#>  6              25    3755132
#>  7              30    4169859
#>  8              35    4694734
#>  9              40    4655093
#> 10              45    3989175
#> # … with 11 more rows
get_polymod_population("Italy") 
#> # A tibble: 21 × 2
#>    lower.age.limit population
#>              <int>      <dbl>
#>  1               0    2758749
#>  2               5    2712790
#>  3              10    2815246
#>  4              15    2917476
#>  5              20    3168918
#>  6              25    3786584
#>  7              30    4553115
#>  8              35    4797005
#>  9              40    4701976
#> 10              45    4136438
#> # … with 11 more rows