Example dataset with information on age based vaccination coverage, acquisition and transmission

data frame with information on vaccine coverage, efficacy of acquisition/susceptibility and efficacy of transmission/infectiousness for the ordered age groups from lowest to highest of the next generation matrix.

vaccination_effect_example_data

Format

A data frame with 17 rows and 4 variables

age_band

character. age bands: 0-4,5-11, 12-15, 16-19, 20-24, etc

coverage

example vaccination coverage, between 0-1

acquisition

example acquisition coverage, between 0-1

transmission

example transmission coverage, between 0-1