Add features required for modelling to the dataset

This function adds three main groups of features to the data. It is used internally in fit_single_contact_model() and predict_contacts_1y(). It requires columns named age_to and age_from. The three types of features it adds are described below:

Population distribution of contact ages from the function add_population_age_to(), which requires a column called "age_to" representing the age of the person who had contact. It creates a column called pop_age_to. add_population_age_to() takes an extra argument for population, which defaults to get_polymod_population(), but needs to be a data frame with columns, lower.age.limit, and population.

School work participation, which is from the function add_school_work_participation(). This requires columns age_to and age_from, but will operate on any column starting with age and adds columns: school_probability, work_probability, school_year_probability, and school_weighted_pop_fraction.

Offset is added on to the data using add_offset(). This requires variables school_weighted_pop_fraction (from add_school_work_participation()) and pop_age_to (from add_school_work_participation()). It adds two columns, log_contactable_population_school, and log_contactable_population.

add_modelling_features(contact_data, ...)

Arguments

contact_data

contact data with columns age_to and age_from

...

extra dots passed to population argument of add_population_age_to()

Value

data frame with 11 extra columns - the contents of contact_data, plus: pop_age_to, school_fraction_age_from, work_fraction_age_from, school_fraction_age_to, work_fraction_age_to, school_probability, work_probability, school_year_probability, school_weighted_pop_fraction, log_contactable_population_school, and log_contactable_population.

Examples

age_min <- 10
age_max <- 15
all_ages <- age_min:age_max
library(tidyr)
example_df <- expand_grid(
   age_from = all_ages,
   age_to = all_ages,
   )
add_modelling_features(example_df) 
#> # A tibble: 36 × 14
#>    age_from age_to pop_age_to intergen…¹ schoo…² work_…³ schoo…⁴ work_…⁵ schoo…⁶
#>       <int>  <int>      <dbl>      <int>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1       10     10      0.161          0       1    0.05       1    0.05       1
#>  2       10     11      0.163          1       1    0.05       1    0.05       1
#>  3       10     12      0.165          2       1    0.05       1    0.2        1
#>  4       10     13      0.168          3       1    0.05       1    0.2        1
#>  5       10     14      0.170          4       1    0.05       1    0.2        1
#>  6       10     15      0.173          5       1    0.05       1    0.2        1
#>  7       11     10      0.161          1       1    0.05       1    0.05       1
#>  8       11     11      0.163          0       1    0.05       1    0.05       1
#>  9       11     12      0.165          1       1    0.05       1    0.2        1
#> 10       11     13      0.168          2       1    0.05       1    0.2        1
#> # … with 26 more rows, 5 more variables: work_probability <dbl>,
#> #   school_year_probability <dbl>, school_weighted_pop_fraction <dbl>,
#> #   log_contactable_population_school <dbl>, log_contactable_population <dbl>,
#> #   and abbreviated variable names ¹​intergenerational,
#> #   ²​school_fraction_age_from, ³​work_fraction_age_from,
#> #   ⁴​school_fraction_age_to, ⁵​work_fraction_age_to, ⁶​school_probability